Moving Forms, Inbox, and UserSection to routes as Pages; Updating imports
