--- src/components/Users/index.js
@@ -1,4 +1,0 @@
-import UserSection from './UserSection';
-import UsersList from './UsersList';
-
-export { UserSection, UsersList };
